requirement.txt
streamlit
Image
pickle
numpy
pandas
Surv
RandomSurvivalForest
Pillow